+++ kamehouse-commons-core/src/main/java/com/nicobrest/kamehouse/commons/annotations/Masked.java
@@ -1,0 +1,15 @@
+package com.nicobrest.kamehouse.commons.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Marker annotation to mask fields that shouldn't be printed in toString() using JsonUtils.
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface Masked {
+
+}