+++ changelog.md
@@ -5,6 +5,7 @@
 * Updated jquery to it's latest version
 * Fixed websockets for kame.com and going through vm-ubuntu.com
 * Added fallback synchronization of vlc player through http calls when the websockets are disconnected
+* Added a @Masked annotation to mask hidden fields on toString() calls
 
 #### v6.05